+++ crates/engine/tests/templates/access_rules/src/lib.rs
@@ -456,7 +456,9 @@ mod access_rules_template {
         pub fn set_component_owner_rule_then_get_owner_proof(&mut self, owner_rule: SubstateOwnerRule) {
             let manager = ComponentManager::get(CallerContext::current_component_address());
             manager.set_owner_rule(owner_rule);
-            manager.get_owner_proof().drop();
+            if let Some(proof) = manager.get_owner_proof() {
+                proof.drop();
+            }
         }
 
         pub fn update_tokens_access_rule(&mut self, action: ResourceAuthAction, new_rule: AccessRule) {